Refactor ConnectionManager out into connection.py
